Add blind signatures
On the blockchain the blind signature will act as a way to verify that the user is eligible to vote. The election authority will verify the user's personal data and blindly sign the user's blinded message which will be the public key of the digital signature that the user will be using on the blockchain.