Add login screen to client browser extension (#57)
* Add login screen to client browser extension

First part of the client (which is implemented as a browser extension) is the login screen. This uses AEAD encryption under the hood to store all the keys in the browsers local storage. Besides that, registering a new account also generates a new signing key. The storage still needs to be implemented though.