Fix undeterministic ports in testing environment (#84)
Previously the testing environment was using dynamic ports on every run, which were chosen from a range at random. This means that, if the environment is restarted, the election config, which is saved in the client's browser memory, becomes invalid. Also continuing work on renaming blockchain config to election config. Note that disabling dynamic port allocation has it's downsides, but the tradeoff is worth it at this time of development.