Add frontend UI to mock authority (#85)
* Add frontend UI to mock authority

Add a frontend UI to the mock authority to streamline the demonstration and manual testing of the voting system. Also restructure the mock authority into it's own subcrate with a build script for the frontend.

* Fix CI issues

`ring` version `0.17.9` seems to break `trunk` builds for some reason, so force it to `0.17.8` for now. Also disable building `mock_authority` `frontend` for now, because `trunk` isn't installed on CI yet. Installing it on MacOS and Windows`might be tricky tho.